+++ coloraide/color.py
@@ -1226,35 +1226,13 @@ class Color(metaclass=ColorMeta):
 
         return [c.convert(out_space, in_place=True) for c in harmonies.harmonize(self, name, space, **kwargs)]
 
-    @deprecated('Please use the class method Color.layer([source, backdrop])')
-    def compose(
-        self,
-        backdrop: ColorInput | Sequence[ColorInput],
-        *,
-        blend: str | bool = True,
-        operator: str | bool = True,
-        space: str | None = None,
-        out_space: str | None = None,
-        in_place: bool = False
-    ) -> Self:  # pragma: no cover
-        """Blend colors using the specified blend mode."""
-
-        if not isinstance(backdrop, str) and isinstance(backdrop, Sequence):
-            colors = [self._handle_color_input(c) for c in backdrop]
-            colors.insert(0, self)
-        else:
-            colors = [self, self._handle_color_input(backdrop)]
-
-        color = compositing.compose(type(self), colors, blend, operator, space, out_space)
-        return self._hotswap(color) if in_place else color
-
     @classmethod
     def layer(
         cls,
         colors: Sequence[ColorInput],
         *,
-        blend: str | bool = True,
-        operator: str | bool = True,
+        blend: str | bool = 'normal',
+        operator: str | bool = 'source-over',
         space: str | None = None,
         out_space: str | None = None
     ) -> Self:
@@ -1264,7 +1242,19 @@ class Color(metaclass=ColorMeta):
         Colors are overlaid on each other with left being the top of the stack and right being the bottom of the stack.
         """
 
-        return compositing.compose(cls, colors, blend, operator, space, out_space)
+        if isinstance(blend, bool):
+            b = 'normal' if blend else None
+            warn_deprecated("The use of boolean values for 'blend' has been deprecated, please use a specific mode")
+        else:
+            b = blend
+
+        if isinstance(operator, bool):
+            o = 'source-over' if operator else None
+            warn_deprecated("The use of boolean values for 'blend' has been deprecated, please use a specific mode")
+        else:
+            o = operator
+
+        return compositing.compose(cls, colors, b, o, space, out_space)
 
     def delta_e(
         self,